v0.02
Finished up the basics of the plugin to make it work for now. As of yet
it does not check the signs nor if it was made by the admin. For
instance if the user knows the permission sets they can make free signs
if there able to place signs and we need to stop this.